tratando erros para id inválidos

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -17,6 +17,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos")
@@ -51,22 +52,32 @@ public class TopicosController {
      }
 
      @GetMapping("/{id}")
-     public DetalhesDoTopicoDTO detalhar(@PathVariable Long id) {
-        return new DetalhesDoTopicoDTO(topicoRepository.getById(id));
+     public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable Long id) {
+         Optional<Topico> topico = topicoRepository.findById(id);
+        if (topico.isPresent()){
+            return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+        }
+        return ResponseEntity.notFound().build();
      }
 
      @PutMapping("/{id}") @Transactional
      public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-        Topico topico = form.atualizar(id, topicoRepository);
-
-        return ResponseEntity.ok(new TopicoDTO(topico));
+         Optional<Topico> topicoValid = topicoRepository.findById(id);
+         if (topicoValid.isPresent()){
+             Topico topico = form.atualizar(id, topicoRepository);
+             return ResponseEntity.ok(new TopicoDTO(topico));
+         }
+         return ResponseEntity.notFound().build();
      }
 
      @DeleteMapping("/{id}") @Transactional
      public ResponseEntity remover(@PathVariable Long id){
-        topicoRepository.deleteById(id);
-
-        return ResponseEntity.ok().build();
+         Optional<Topico> topico = topicoRepository.findById(id);
+         if (topico.isPresent()){
+             topicoRepository.deleteById(id);
+             return ResponseEntity.ok().build();
+         }
+         return ResponseEntity.notFound().build();
      }
 
 }
